+++ clibs/gameplay/src/system/hub.cpp
@@ -438,20 +438,20 @@ static std::tuple<size_t, size_t, bool> FindHub(world& w, const hub_mgr::hub_inf
             auto& chestslot = chest::array_at(w, container::index::from(*chest), berth.chest_slot);
 
             auto amount1 = chestslot.amount + chestslot.lock_space;
-            if (((min.index == -1) || (amount1 < min.amount)) && (chestslot.limit > amount1)) {
+            if (((min.index == (size_t)-1) || (amount1 < min.amount)) && (chestslot.limit > amount1)) {
                 min.index = ii;
                 min.amount = amount1;
             }
 
             auto amount2 = chestslot.amount - chestslot.lock_item;
-            if (((max.index == -1) || (amount2 > max.amount)) && (amount2 > 0)) {
+            if (((max.index == (size_t)-1) || (amount2 > max.amount)) && (amount2 > 0)) {
                 max.index = ii;
                 max.amount = amount2;
             }
         }
     }
     bool moveable = false;
-    if(min.index != -1) { 
+    if(min.index != (size_t)-1) { 
         moveable = min.amount + 2 <= max.amount;
     }
     return {min.index, max.index, moveable};
